# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Attendance Management API
  version: 1.0.0
  description: >
    Sistem absensi sekolah berbasis Node.js, Express, Sequelize, RBAC, dan semester.
    Mendukung absensi harian, absensi per jam, validasi, rekap, data master,
    manajemen user, scheduler otomatis, export, file bukti, serta logging aktivitas.
servers:
  - url: http://localhost:3000
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        nisn: { type: string }
        username: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [super_admin, bk, wali_kelas, guru_mapel, perangkat_kelas, siswa]
        nama_lengkap: { type: string }
        kelas_id: { type: integer }
        jurusan_id: { type: integer }
        is_active: { type: boolean }
        last_login:
          type: string
          format: date-time
...
paths:
...
  /api/user:
    get:
      tags: [User]
      summary: List user
      security:
        - BearerAuth: []
      responses:
        200:
          description: List user

    post:
      tags: [User]
      summary: Create user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/User' }
      responses:
        201:
          description: User dibuat
  /api/user/{id}:
    get:
      tags: [User]
      summary: Detail user
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: Detail user

    put:
      tags: [User]
      summary: Update user
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/User' }
      responses:
        200:
          description: User updated

    delete:
      tags: [User]
      summary: Hapus user
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: User deleted


  ##############################################
  # JURUSAN
  ##############################################
...
  /api/kelas:
    get:
      tags: [Kelas]
      summary: List kelas
      security:
        - BearerAuth: []
    post:
      tags: [Kelas]
      summary: Create kelas
      security:
        - BearerAuth: []
  /api/kelas/{id}:
    get:
      tags: [Kelas]
      summary: Detail kelas
      security:
        - BearerAuth: []
    put:
      tags: [Kelas]
      summary: Update kelas
      security:
        - BearerAuth: []
    delete:
      tags: [Kelas]
      summary: Delete kelas
      security:
        - BearerAuth: []


  ##############################################
  # MAPEL
  ##############################################
  /api/mapel:
    get:
      tags: [Mapel]
      summary: List mapel
      security:
        - BearerAuth: []
    post:
      tags: [Mapel]
      summary: Create mapel
      security:
        - BearerAuth: []
...